+++ google/src/main/scala/org/broadinstitute/dsde/workbench/google/GooglePubSubDAO.scala
@@ -21,7 +21,8 @@ object GooglePubSubDAO {
   case object MessageNotHandled extends HandledStatus
   case object NoMessage extends HandledStatus
 
-  case class PubSubMessage(ackId: String, contents: String)
+  case class PubSubMessage(ackId: String, contents: String, attributes: Map[String, String] = Map.empty)
+  case class MessageRequest(text: String, attributes: Map[String, String] = Map.empty)
 }
 
 trait GooglePubSubDAO {
@@ -35,15 +36,25 @@ trait GooglePubSubDAO {
 
   def setTopicIamPermissions(topicName: String, permissions: Map[WorkbenchEmail, String]): Future[Unit]
 
-  def createSubscription(topicName: String, subscriptionName: String): Future[Boolean]
+  def createSubscription(topicName: String, subscriptionName: String, ackDeadlineSeconds: Option[Int]): Future[Boolean]
 
   def deleteSubscription(subscriptionName: String): Future[Boolean]
 
-  def publishMessages(topicName: String, messages: scala.collection.Seq[String]): Future[Unit]
+  def publishMessages(topicName: String, messages: scala.collection.Seq[MessageRequest]): Future[Unit]
 
   def acknowledgeMessages(subscriptionName: String, messages: scala.collection.Seq[PubSubMessage]): Future[Unit]
 
   def acknowledgeMessagesById(subscriptionName: String, ackIds: scala.collection.Seq[String]): Future[Unit]
+
+  def extendDeadline(subscriptionName: String,
+                     messages: scala.collection.Seq[PubSubMessage],
+                     extendDeadlineBySeconds: Int
+  ): Future[Unit]
+
+  def extendDeadlineById(subscriptionName: String,
+                         ackIds: scala.collection.Seq[String],
+                         extendDeadlineBySeconds: Int
+  ): Future[Unit]
 
   def pullMessages(subscriptionName: String, maxMessages: Int): Future[scala.collection.Seq[PubSubMessage]]