+++ google/src/test/scala/org/broadinstitute/dsde/workbench/google/mock/MockGoogleStorageDAO.scala
@@ -2,13 +2,21 @@ package org.broadinstitute.dsde.workbench.google.mock
 
 import java.io.{ByteArrayInputStream, ByteArrayOutputStream, File}
 import java.nio.file.Files
-
 import com.google.api.client.util.IOUtils
-import com.google.api.services.storage.model.{Bucket, BucketAccessControls, ObjectAccessControls}
+import com.google.api.services.storage.model.{
+  Bucket,
+  BucketAccessControls,
+  ObjectAccessControls,
+  Policy => BucketPolicy
+}
 import org.broadinstitute.dsde.workbench.google.GoogleStorageDAO
+import org.broadinstitute.dsde.workbench.model.WorkbenchEmail
 import org.broadinstitute.dsde.workbench.model.google.GcsLifecycleTypes.{Delete, GcsLifecycleType}
 import org.broadinstitute.dsde.workbench.model.google.GcsRoles.GcsRole
 import org.broadinstitute.dsde.workbench.model.google.{GcsBucketName, GcsEntity, GcsObjectName, GoogleProject}
+import org.broadinstitute.dsde.workbench.google.HttpGoogleStorageDAO._
+import org.broadinstitute.dsde.workbench.model.google.iam.IamMemberTypes.IamMemberType
+import org.broadinstitute.dsde.workbench.model.google.iam.{Expr, Policy}
 
 import scala.collection.concurrent.TrieMap
 import scala.concurrent.{ExecutionContext, Future}
@@ -180,4 +188,24 @@ class MockGoogleStorageDAO(implicit val executionContext: ExecutionContext) exte
     Future.successful(new ObjectAccessControls())
 
   override def setRequesterPays(bucketName: GcsBucketName, requesterPays: Boolean): Future[Unit] = Future.successful(())
+
+  override def addIamRoles(bucketName: GcsBucketName,
+                           userEmail: WorkbenchEmail,
+                           memberType: IamMemberType,
+                           rolesToAdd: Set[String],
+                           retryIfGroupDoesNotExist: Boolean = false,
+                           condition: Option[Expr] = None,
+                           userProject: Option[GoogleProject]
+  ): Future[Boolean] = Future.successful(false)
+
+  override def removeIamRoles(bucketName: GcsBucketName,
+                              userEmail: WorkbenchEmail,
+                              memberType: IamMemberType,
+                              rolesToRemove: Set[String],
+                              retryIfGroupDoesNotExist: Boolean = false,
+                              userProject: Option[GoogleProject]
+  ): Future[Boolean] = Future.successful(false)
+
+  override def getBucketPolicy(bucketName: GcsBucketName, userProject: Option[GoogleProject]): Future[BucketPolicy] =
+    Future.successful(Policy(Set.empty, ""))
 }